Multiple Atoms tab now shows legend and nucleus chooser panels
